refactoring CardController before migrating to IGListKit
